feat: create commander.js cli
The image / video to ascii converter can now run fully in javascript
through a commander.js cli. There are various flags to change the
programs behavior but much of them are not yet implemented. The program
ensures that all files created that are not needed in the final output
are cleaned up. The output file will also never overwrite an existing
file unless a user passes the program the [--force] option flag.