refactor: use `string[]` instead of `string` to represent rows
This change will negatively impact performance but is necesarry for the
[-s|--space] option to work proerly for when a user specifies a space
replacer that is more then one character long (usually escape sequences)
such as "&nbsp;". A value like that would change how the string is
indexed but if they are arrays instead it will not change the length of
the array as the string is seens as a single unit / entry in the array.