fix: make sure photo input is previewable
The issue was that the photo was only being printed to an alternate
buffer then that buffer would immediately close. This made it seem like
the ascii photo was never being printed but in reality it was just
becasue there was no work to keep doing in the node event loop.